Add duration funcs, update time prompt
Add `toggl-parse-duration` for parsing a duration string into an
integer number of seconds.

Add `toggl-prompt-duration` for prompting a user to enter a duration.

Update `toggl-api-duration` to handle more types of inputs, and to
make use of `toggl-parse-duration`.

Add `toggl-prompt-time` for prompting a user to enter a timestamp.

Fix `toggl-api-timestamp` to not mess up already well-formatted
timestamps.